requirements.txt
requests
python-dotenv
pytest
mock
flask
flask-ask